修复3.11日告警

+++ wmserver/src/window_controller.cpp
@@ -300,7 +300,7 @@ void WindowController::ProcessDisplayChange(DisplayId displayId, DisplayStateCha
 
             // Remove 'sysBarWinId_' after SystemUI resize 'systembar'
             uint32_t width = static_cast<uint32_t>(displayInfo->GetWidth());
-            uint32_t height = displayInfo->GetHeight() * SYSTEM_BAR_HEIGHT_RATIO;
+            uint32_t height = static_cast<uint32_t>(displayInfo->GetHeight() * SYSTEM_BAR_HEIGHT_RATIO);
             Rect newRect = { 0, 0, width, height };
             ResizeRect(sysBarWinId_[WindowType::WINDOW_TYPE_STATUS_BAR], newRect, WindowSizeChangeReason::DRAG);
             newRect = { 0, displayInfo->GetHeight() - height, width, height };
